+++ src/server/game/Server/WorldSession.h
@@ -495,11 +495,11 @@ class WorldSession
         uint16 GetClientBuild() const { return m_ClientBuild; }
 
         /// Return join date as unix timestamp
-        uint32 GetAccountJoinDate() const { return m_AccountJoinDate; }
+        uint64 GetAccountJoinDate() const { return m_AccountJoinDate; }
 
         /// Set join date as unix timestamp
         /// @p_JoinDate : unix timestamp of the account creation
-        void SetAccountJoinDate(uint32 p_JoinDate) { m_AccountJoinDate = p_JoinDate; }
+        void SetAccountJoinDate(uint64 p_JoinDate) { m_AccountJoinDate = p_JoinDate; }
 
         time_t GetLoginTime() const { return m_LoginTime; }
 
@@ -1290,7 +1290,7 @@ class WorldSession
 
         uint16 m_ClientBuild;
 
-        uint32 m_AccountJoinDate;
+        uint64 m_AccountJoinDate;
 
         //////////////////////////////////////////////////////////////////////////
         /// Premium
